MC-37324: Add fallback to 'patch' command when 'git' command is not available
- Remove redundant exceptions
